requirements.txt
mysql-connector-python~=8.0.30
requests~=2.28.1